Major changes to "Player", "Controller", "World" and "Render3D".
Player: Holds and manages it's own potential collidables and collision handler. Also it returns true or false at the end of the function "tick" to indicate if the player is alive or not.
Controller: Removed use of old camera function "rotateSideways" and added function that creates a projectile if the left mouse button is clicked.
World: Keeps track of projectiles in the world and uses "Render3D" to render them. Furthermore it's not managing the players potential collidables.
Render3D: Render projectiles in addition to environment.